+++ CHANGELOG.md
@@ -13,7 +13,7 @@
     - New modules, evolution of 8FACE and 8FACEx2 (#63 #76 #144 #154 #157 #158 #160 #162)
 - Module [DIRT](./docs/Dirt.md)
     - New module, crosstalk and noise for polyphonic cables
-- Module [MIDI-CAT MAP](./docs/MidiCat.md#map-expander)
+- Module [MIDI-CAT CTX](./docs/MidiCat.md#ctx-expander)
     - New expander-module for MIDI-CAT, helper for mapping parameters by context menu (#232)
 - Module [MIDI-PLUG](./docs/MidiPlug.md)
     - New module, a virtual MIDI merger and splitter